splite to multible livewire componints
